Refactor: Rename workspace to playlist (files)
