+++ src/gui/actions.rs
@@ -16,7 +16,7 @@ use super::{
     modals::file_dialogs,
     GuiState,
 };
-use crate::player::{workspace::enums::FileListMode, Player};
+use crate::player::{playlist::enums::FileListMode, Player};
 
 // --- Common File Actions --- //
