+++ src/gui/playlist_fonts.rs
@@ -8,8 +8,8 @@ use super::{
     GuiState, TBL_ROW_H,
 };
 use crate::player::{
+    playlist::{enums::FileListMode, font_meta::FontMeta},
     soundfont_list::FontSort,
-    workspace::{enums::FileListMode, font_meta::FontMeta},
     Player,
 };
