+++ src/player/playlist/enums.rs
