+++ src/player/playlist/error.rs
