+++ src/player/playlist/font_meta.rs
@@ -120,7 +120,7 @@ impl TryFrom<&serde_json::Value> for FontMeta {
 mod tests {
 
     use super::*;
-    use crate::player::workspace::Playlist;
+    use crate::player::playlist::Playlist;
     use serde_json::Value;
 
     fn run_serialize(playlist: Playlist) -> Playlist {
